feat: update expense amounts to integers and enhance expense item display

+++ lib/models/expense.dart
@@ -15,7 +15,7 @@ class Expense {
 
   final String id;
   final String title;
-  final double amount;
+  final int amount;
   final DateTime date;
   final Category category;
 }
